+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.2.1
 - removed duplicate CLI functions
+- fixed Maestro writing error
   
 ## 1.2.0
 - added Maestro class that converts .mae/.maegz files to .cif